Setup trpc context and added manufacturer table
This also has some initial implementation of signing in with username and password.
Before I could continue with that, I pivoted to learning trpc and prisma.

context.ts in src/server was the first task and was coded/revised using the documentation provided by trpc.
This was to accomodate protectedProcedure for enforcing user authentication for some routes.
A manufacturer router/table has been created in src/server/routers and in prisma, respectively, for learning the flow of getting and setting data.