Finished implementing credentials authentication
This implementation is from the perspective of an amateur developer. A look into how its done and revising it is worth doing in the future.

https://github.com/FranciscoMendes10866/next-auth-trpc-prisma/blob/main/src/pages/index.tsx
The link above was used as reference on how to query the database to check for users in the authorize function in [...nextauth].js.
At first, I thought it was necessary to create a trpc user route for checking credentials. However, since useQuery was the only way
for me to use trpc, I couldn't use it in the authorize function since the authorize function is not a react component. To solve this,
I skipped trpc and decided to query the database in the authorize function. I do not know if this okay since I am not entirely sure if [...nextauth].ts
is part of the backend or frontend.
*Might revist this in the future.*

Installed react-hook-form for managing signin and signup pages.

Next:
- Fix the redirect logic when signing up and signing in.