Learned how to set custom layouts
I was previously confused on how _app.tsx worked. After analyzing it, the reason for its existence is for making custom layouts for pages.
If a getLayout function exists in a page component, the return value from it will be the wrapper used for the page instead of DefaultLayout.tsx