Implemented preliminary design and strengthened account passwords
Implemented the figma design using tailwind and minimal sass.
The website is designed to be mobile first so that scaling it up to bigger screens is easier.

Added salting to password hashes to strengthen security.
NOTE: This implementation is very primitive and should be improved upon using the following as an example:
```
export const hashPass = async (password: string): Promise<string> => {
  const salt = await bcrypt.genSalt(saltEnumeration);
  const hashedPassword = await bcrypt.hash(password, salt);
  return hashedPassword;
};
```

Misc:
Stopped localhost:5555 being opened when running pnpm dx