Mostly done with the cars page
It is responsive to all screen widths. The only thing left for it is the correct mapping of the logos to their respective manufacturer.
I have it coded in a way that it assumes that the logos have a square aspect ratio.
https://cloudinary.com/blog/upload-and-display-images-using-prisma
The link above should aid in solving such issue.