Finished styling on manufacturer/car models page
The navigation towards tuning sheets have been reefactored.
The previous implementation is in pages/tmp for reference.
The page still needs to be connected to the list of tuning sheets page, i.e. after clicking on a specific car model, the user is redirect to a list of tuning sheets for  that car.
Also still needs for manufacturers logos and car model images to be properly stored and displayed