Added drag and drop image func. for upload
Followed this tutorial: https://www.section.io/engineering-education/nextjs-dnd-file-upload/

Still needs to be able to send images to the backend. The styling also still needs to be done.
Once the images are in the backend, I hope to parse the text within the images using tesseract.js so that the tuning sheets are made.