Added manufacturer logos
The tailwind for the manufacturer logos have been revamped to accomodate the varying aspect ratios.
A seed has been created for the database where each manufacturer has their image property set to a cloudinary image url.

The idea of using tesseract is scrapped due to its complexity and difficulty to use. Adding tuning sheets will be done in a different way.
The remaining implementation will still be available in the repository for a possible future use.