First implementation of Lobbies
My idea of lobbies are now implemented in the database schema.

My vision is that each lobby pulls from a lobby preset(LobbySettings) that is already exisiting or new from yourself or from someone else.
This way, same races can be run by different people on different times.
These settings included the cars (and their tuning sheets), tracks, and other settings for lobbies that want more flexibility.
Notably, players in a lobby are stored in a list (usersInLobbies) where the host has the attribute isHost set to true.
I am not sure if I should do this or set a specific property for the host.

The relationship between LobbySettings and CarModels is explicit. This is done so that a tuningSheet can be optionally associated with an allowed car in the lobby within the relation table. I am not sure if I should do it like this or create two separate many-to-many relationships with CarModel and TuningSheet.

Schema todos:
- Add more lobby tags
- Possibly add more LobbySettings properties for more customizability.