Started on create lobby page
To create a lobby, the user will need to press the create button on the lobby page. They can then either choose to create a new preset or an existing preset. choosing either will redirect them to a page that allows them to change the details and specific of the lobby. The difference between the two is that the latter will have fields prefilled. This page is what I am currently working on.

To do:
- Selecting multiple cars using the CarSelect component.
The CarSelect component is all the code from the cars page, where one picks on a manufacturer, then a car model to be redirected to a page of all the car sheets of such model. Instead of this functionality, I want the user to be able to select multiple cars. The issue I have is keeping both of these implementations on the same component. I find this difficult because the car models are Link components, and I think to be able to select multiple cars you would need them to be div components with an onClick attribute that changes a state.

- Selecting the car sheet for each of the selected cars allowed.
I am thinking of providing a dropdown of tuning sheets for each selected car. The number of dropdowns will change depending on how many cars you select. This will definitely be need to be redone because it does not make it easy for users to find tuning sheets.

Code refactored:
CarSelect.tsx - instead of using url slugs for displaying which manufacturer car models, useState is used.
LobbyCard.tsx - has the appropriate props for filling the text with real data from the database.