Can now select a tuning sheet for each car model
After selecting multiple cars allowed for the lobby, buttons appear corresponding to each selected car. Each button will allow the user to select a tuning sheet for a particular car. There needs to be a way for the user to filter which tuning sheets they want as pressing a button will show them all the tuning sheets. I am thinking of implementing a search bar and filters in the future.

A modal component has been implemented using this article: https://blog.bitsrc.io/simple-accessible-modal-in-react-typescript-and-tailwind-3296704a985a

Each manufacturer has a model to fix some styling issues.

MultipleCarSelect.tsx -> MultipleModelSelect.tsx
This was done for semantics. There are two interface in this file that cause it be coupled with src/pages/lobbies/create/[lobbyPresetId].tsx.

Widget.tsx now has an X button that can opted in by defining a onClose function. This was for using it in conjuction with Modal.tsx and useState.