Can now add tracks to lobby settings using data from datbase
Previously, selecting tracks was done using a drop down with filler value. Now, the tracks (TrackLayouts) are pull from the backend and displayed as checkboxes. Users can now choose multiple tracks and submit it along with the other settings seemlessly. The selection is tracked with a useState hook.

Tags and GR rating options are also pulled from the database. The text still needs to be change to make it look better for the users. i.e. change streetTires to "Street Tires" and gr1 to "Gr. 1".