Finished making all options for creating lobbies
The issue now is passing the values to the backend. I cannot seem to pass it with the current implementation due to zod conflictions. The form data is unable to meet the "title", "description", and other field criterias of the zod object specified in the trpc route. I will now try to implement react hook form.