Can now submit lobby settings to the backend.
With the use of react-hook-form and 5 useEffects, the information filled by user is validated and can now be submitted to the backend safetely.
I feel that there could be a better way of handling the useState data used for useForm without the 5 useEffects. I will probably comeback to this in the future.

Now the main task should be filling the database with the information. After that, I could start on rendering the lobby settings in src/pages/lobbies/create/index.tsx.