Premade lobby settings now display on the website
These lobby settings are pulled from the data base and rendered on src/pages/lobbies/create/index.tsx. Clicking on one will redirect the user to src/pages/lobbies/create/[lobbySettingId].tsx , the edit lobby settings page, where the fields are prefilled.

Notably, a trpc route is not used for prefilling fields in the edit lobby settings page. Rather, a function located in the lobby router file is imported and used in the getServerSideProps function of the page. This is so that the fields are filled when the page is pre-rendered.

I am not the most knowledgeable in SSR but I believe this is still a better way than using another useQuery after loading the page and then filling in the fields.

Database changes:
- added a createdAt field for model LobbySettings.

- LobbyCard.tsx: refactored to dynamically display data. Also have been commented out in pages where no props are passed in.
- [lobbySettingId].tsx: provided default values for various fields.
- src/pages/lobbies/create/index.tsx: Implemented the logic from the post example page here so that the entries are loaded 5 at a time.