Can now update specific lobby settings
This is a decent implementation as it works but not to the best standards. When updating it passes all the values from the frontend to the backend,
including values that have not changed. Then an update on the specific lobby setting using all the values.
I imagine this might not be the most efficient so I plan on improving this with the next commit.

I am currently trying on using dirtyFields in useForm but I cannot make it work with the fields using useState (e.g. selectedModels, tracks, tags, etc.).
It also does not work when the page loads. I tested this by console logging dirtyFields after interacting with the text fields such as title and description.
I have used this for reference: https://codesandbox.io/s/react-hook-form-submit-only-dirty-fields-ol5d2?file=/src/index.tsx:2049-2118
Some implementation has been done in src/utils/misc.ts where a getDirtyValues function exists.

Other changes:
- Button has now been split into Button and ButtonLink