lobby settings are updated only with the necessary changes
Rather than passing all the data to the backend and updating an all the attributes in an entry, it is done now with only the changed fields.
This was done using React Hook Form's dirty fields where differences from the defaultValues are tracked. Only dirty fields are passed to the backend.
The dirty fields were also used to mark/color fields that have been changed by the user.
Updates can only be done if the original curator of the lobby settings wants to change any of the options.

Other Changes:
- custom Button className prop now binds to the inner button html element.