Correct lobby settings values are now being updated.
Formerly, array fields such as cars, tags, tracks, etc. were being marked as dirty despite having the same content as their defaultValues counterpart.
After further research, it was discovered that the order of these array fields were observed by RHF.
["supra", "r34"] is different from ["r34", "supra"].

To solve this, I implemented sorting for cars and tracks according to their names in ascending order. These two fields are sorted when pulled
from the database, and whenever the user selects/deselects one of their options.
As for the grRating field and the tags field, they are much easier to deal with as the sort is already built in.

The reason I had to do more for cars and tracks was due to them being objects rather than strings.

Other changes:
- The original curator for each lobby settings now displays on the form.
- When prefilling the lobby settings form, the creator id is now also pulled.
- when a "save" is done to lobby settings, the creator id is now also pushed to the backend.
In conjuction with the lobby setting id, It is used to find the specific lobby setting to update.
I believe this will increase security against other users updating lobby settings not belonging to them.
- sortObjectByStringCallback function has been implemented in src/utils/misc.ts. This is mainly used for sorting tracks and cars when user select/deselect.